refactor: extract component demo data to separate files
Move inline demo/default data from components to dedicated demo/data.ts
files in each category folder. This improves code organization and makes
components more maintainable by separating presentation logic from
sample data.

Categories updated:
- events: demoEvent, demoEvents, demoEventDetails
- blogging: demoPost, demoPosts
- messaging: demoMessages
- list: demoProducts
- payment: demoOrderItems, demoOrderData
- miscellaneous: demoOptions
