+++ packages/manifest-ui/registry.json
@@ -85,7 +85,7 @@
     },
     {
       "name": "order-summary",
-      "version": "1.0.0",
+      "version": "1.0.1",
       "type": "registry:component",
       "title": "Order Summary",
       "description": "An order summary block displaying items, subtotal, shipping, tax, discounts, and total.",
@@ -197,7 +197,7 @@
     },
     {
       "name": "product-list",
-      "version": "2.0.1",
+      "version": "2.0.2",
       "type": "registry:component",
       "title": "Product List",
       "description": "Product list with list, grid, carousel, and picker variants.",
@@ -213,7 +213,7 @@
     },
     {
       "name": "option-list",
-      "version": "2.0.0",
+      "version": "2.0.1",
       "type": "registry:component",
       "title": "Option List",
       "description": "Tag-style option selector with single or multiple selection modes.",
@@ -341,7 +341,7 @@
     },
     {
       "name": "post-card",
-      "version": "2.0.0",
+      "version": "2.0.1",
       "type": "registry:component",
       "title": "Post Card",
       "description": "Post card with image, title, excerpt, author info and read more button. Supports default, compact, horizontal, and covered variants. Includes post-detail for fullscreen view.",
@@ -361,7 +361,7 @@
     },
     {
       "name": "post-list",
-      "version": "2.0.1",
+      "version": "2.0.2",
       "type": "registry:component",
       "title": "Post List",
       "description": "Post list with list, grid, carousel, and fullwidth variants. Fullwidth mode shows paginated posts.",
@@ -385,7 +385,7 @@
     },
     {
       "name": "post-detail",
-      "version": "2.0.0",
+      "version": "2.0.1",
       "type": "registry:component",
       "title": "Post Detail",
       "description": "Full post view with cover image, author info, content, tags and related posts section.",
@@ -437,7 +437,7 @@
     },
     {
       "name": "chat-conversation",
-      "version": "4.0.1",
+      "version": "4.0.2",
       "type": "registry:component",
       "title": "Chat Conversation",
       "description": "Full chat conversation component with multiple message types.",
@@ -537,7 +537,7 @@
     },
     {
       "name": "event-card",
-      "version": "5.0.1",
+      "version": "5.0.2",
       "type": "registry:component",
       "title": "Event Card",
       "description": "Display event information with multiple layouts. Supports events with images, signals, vibe tags, and display-formatted date/time. Entire card is clickable.",
@@ -557,7 +557,7 @@
     },
     {
       "name": "event-list",
-      "version": "6.0.1",
+      "version": "6.0.2",
       "type": "registry:component",
       "title": "Event List",
       "description": "Display events in grid, list, or carousel layouts. Fullscreen mode shows interactive split-screen map with Leaflet and animated filter panel.",
@@ -577,7 +577,7 @@
     },
     {
       "name": "event-detail",
-      "version": "3.0.1",
+      "version": "3.0.2",
       "type": "registry:component",
       "title": "Event Detail",
       "description": "Full event detail view with organizer info, interactive map, policies, and ticket purchase. Fullwidth mode only.",
