+++ packages/manifest-ui/registry/list/product-list.tsx
@@ -4,6 +4,7 @@ import { Button } from '@/components/ui/button'
 import { cn } from '@/lib/utils'
 import { Check, ChevronLeft, ChevronRight, ShoppingCart, Star } from 'lucide-react'
 import { useCallback, useState } from 'react'
+import { demoProducts } from './demo/data'
 
 /*
  * ProductList Component - ChatGPT UI Guidelines Compliant
@@ -46,60 +47,6 @@ export interface ProductListProps {
     selectedProductIndex?: number
   }
 }
-
-const defaultProducts: Product[] = [
-  {
-    name: "Air Force 1 '07",
-    description: 'Nike',
-    price: 119,
-    image: '/demo/shoe-1.png',
-    rating: 4.9,
-    badge: 'New',
-    inStock: true
-  },
-  {
-    name: 'Air Max 90',
-    description: 'Nike',
-    price: 140,
-    image: '/demo/shoe-2.png',
-    rating: 4.8,
-    inStock: true
-  },
-  {
-    name: 'Air Max Plus',
-    description: 'Nike',
-    price: 170,
-    originalPrice: 190,
-    image: '/demo/shoe-4.png',
-    rating: 4.7,
-    badge: '-10%',
-    inStock: true
-  },
-  {
-    name: 'Dunk Low',
-    description: 'Nike',
-    price: 115,
-    image: '/demo/shoe-3.png',
-    rating: 4.6,
-    inStock: true
-  },
-  {
-    name: 'Jordan 1 Low',
-    description: 'Nike',
-    price: 135,
-    image: '/demo/shoe-1.png',
-    rating: 4.8,
-    inStock: true
-  },
-  {
-    name: 'Blazer Mid',
-    description: 'Nike',
-    price: 105,
-    image: '/demo/shoe-2.png',
-    rating: 4.5,
-    inStock: true
-  }
-]
 
 // Horizontal card for list variant
 function ProductHorizontalCard({
@@ -787,7 +734,7 @@ function PickerVariant({
 }
 
 export function ProductList({ data, actions, appearance, control }: ProductListProps) {
-  const { products = defaultProducts } = data ?? {}
+  const { products = demoProducts } = data ?? {}
   const { onSelectProduct, onAddToCart } = actions ?? {}
   const { variant = 'list', currency = 'EUR', columns = 4, buttonLabel } = appearance ?? {}
   const { selectedProductIndex } = control ?? {}
