+++ packages/manifest-ui/registry/miscellaneous/demo/data.ts
@@ -1,0 +1,11 @@
+// Demo data for Miscellaneous category components
+// This file contains sample data used for component previews and documentation
+
+import type { Option } from '../option-list'
+
+// Default options for OptionList
+export const demoOptions: Option[] = [
+  { label: 'Standard shipping', description: '3-5 business days' },
+  { label: 'Express shipping', description: '1-2 business days' },
+  { label: 'Store pickup', description: 'Available in 2h' }
+]
